Atualização Projeto Devsbook Orientado à Objetos

+++ DEVSBOOKOO/dao/PostDaoMysql.php
@@ -27,11 +27,39 @@ class PostDaoMysql implements PostDAO {
     }
 
     public function delete($id, $id_user) {
-        $sql = $this->pdo->prepare("DELETE FROM posts
+        $postLikeDao = new PostLikeDaoMysql($this->pdo);
+        $postCommentDao = new PostCommentDaoMysql($this->pdo);
+
+        // 1. Verificar se o POST existe (Para verificar o tipo)
+        $sql = $this->pdo->prepare("SELECT * FROM posts
         WHERE id = :id AND id_user = :id_user");
         $sql->bindValue(':id', $id);
         $sql->bindValue(':id_user', $id_user);
         $sql->execute();
+
+        if($sql->rowCount() > 0) {
+            $post = $sql->fetch(PDO::FETCH_ASSOC);
+            
+            // 2. Deletar os LIKES e COMMENTS
+            $postLikeDao->deleteFromPost($id);
+            $postCommentDao->deleteFromPost($id);
+            
+            // 3. Deletar a eventual foto (type === photo)
+            if($post['type'] === 'photo') {
+                $img = 'media/uploads/'.$post['body'];
+                if(file_exists($img)) {
+                    unlink($img);
+                }
+            }
+            
+            // 4. Deletar o POST
+            $sql = $this->pdo->prepare("DELETE FROM posts
+            WHERE id = :id AND id_user = :id_user");
+            $sql->bindValue(':id', $id);
+            $sql->bindValue(':id_user', $id_user);
+            $sql->execute();
+
+        }
     }
 
     public function getUserFeed($id_user) {
